+++ packages/stream_chat_localizations/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* Added support for [Norwegian](https://github.com/GetStream/stream-chat-flutter/blob/master/packages/stream_chat_localizations/lib/src/stream_chat_localizations_no.dart) locale.
+
 ## 3.2.0
 
 ✅ Added